+++ sematic/tests/test_future.py
@@ -44,56 +44,46 @@ def test_no_tracking():
     assert foo().resolve(tracking=False) == "foo"
 
 
-def test_getitem():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
-        foo()[0]
-
-
-def test_iter():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
-        a, b = foo()
-
-
 def test_bool():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         if foo():
             pass
 
 
 def test_not():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         if not foo():
             pass
 
 
 def test_add():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         foo() + 1
 
 
 def test_iadd():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         f = foo()
         f += 1
 
 
 def test_mul():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         foo() * 1
 
 
 def test_imul():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         f = foo()
         f *= 1
 
 
 def test_div():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         foo() / 1
 
 
 def test_idiv():
-    with pytest.raises(NotImplementedError, match="docs.sematic.ai"):
+    with pytest.raises(NotImplementedError, match="docs.sematic.dev"):
         f = foo()
         f /= 1