+++ docs/changelog.md
@@ -1,6 +1,16 @@
-* HEAD
+* 0.5.0
+    * [feature] `getitem` support for futures of list, tuples, dictionaries
+    * [feature] `__iter__` support for futures of tuples
+* 0.4.0
+    * [feature] ability to deploy the Sematic API to a cloud instance and run
+      pipelines against it (pipeline still runs locally)
+    * [improvement] Rename `$ sematic credentials` to `$ sematic settings` to be
+      able to store other things than credentials.
+* 0.3.0
     * [feature] Support for Tuple types
     * [feature] Support for Dict types
+    * [feature] `SnowflakeTable` type
+    * [feature] `$ sematic credentials set <app> <var> <value>` CLI command
 * 0.2.0
     * [bugfix] UI scroll issues
     * [bugfix] Dataframe UI previews fails for null/NaN values